fs/config: keep FSConfig's methods with its type
MarshalText and UnmarshalText lived in their own file, away from the
struct they belong to. Move them into config.go, directly below the
FSConfig definition.
